fix+db+docs: F-80 第三批——F-76 甲组「接上真库必然失败」六类已修
承 F-79。本机无 PostgreSQL 16 不能实跑，每条以代码与 DDL 逐字交叉核实，核不实的不改。

已修六类：make_interval 具名实参是 int4 而绑参是 i64=int8、int8→int4 非隐式转换致函数解析 42883，已加
::int，迁移窗口是结构变更唯一放行闸；decode_row 缺 INT4/INT2 分支致落文本解码报 ColumnDecode，而
ColumnDecode 非 Database 错误、sqlstate 为 None、最终映射成 INTERNAL_ERROR 看不出真实原因，已补两分支
归一到 Int64；data_keys 漏 activated_at，DDL 逐字 not null 且无默认值（同表另三个时间列都是 null，只有
它 not null），已补 now()；降级窗口漏 opened_at，DDL not null 无默认而紧邻的 closed_at 有 default
'infinity'——一个有一个没有正是最易看漏的形态；digest 未加 schema 限定，迁移逐字 create extension
pgcrypto schema platform_core 而 public 已删除，已改 platform_core.digest 并反侧确认该 schema 的 USAGE
已授予；两条回填迁移向 FORCE RLS 表写入却不设 app.legal_entity_id，核实链四步走完（三表都调
attach_table_guards → 函数内 if v_has_le then apply_le_rls → apply_le_rls 下 force row level security 加
策略），FORCE 之下连表属主也受约束、WITH CHECK 恒不成立，今天不炸只因法人表为空循环体不执行，两条已各
补 set_config。

未修一条：DbValue::Null 绑 TEXT OID。报告称 uuid 等列 42804，该结论核不实——INSERT 目标列走赋值转换、
text→uuid 有 I/O 转换可能被接受，而 WHERE 比较上下文需隐式转换确会失败，两种上下文结论相反，本机无
PG16 无从判定。已登记转绿判据。

方法记录：反查同类用粗筛捞出 8 个候选，逐个核实后只有 1 个成立——另两个 backfill 的目标表根本没有
legal_entity_id 列，不受 le-RLS 约束。粗筛可定位不可定性。

sqlcheck 0、archcheck 0、真全量 1113/3/5 不变、基线逐面相等，无新回归。本批正确性只由代码与 DDL 逐字
交叉核实支撑，未经真实 PostgreSQL 执行。

+++ crates/adapter/db-pg/src/platform_core/windows.rs
@@ -35,7 +35,10 @@ const LATEST_OPEN_STMT: &str = "select id from platform_core.migration_windows \
 
 const INSERT_WINDOW_STMT: &str = "insert into platform_core.migration_windows \
      (id, state, approval_ref, reason, opened_by, opened_at, expires_at) \
-     values ($1, 'OPEN', $2, $3, $4, now(), now() + make_interval(mins => $5))";
+     values ($1, 'OPEN', $2, $3, $4, now(), now() + make_interval(mins => $5::int))";  // ::int 不可省：绑参是 i64＝int8，
+                                             // 而 make_interval 的具名实参是 int4，
+                                             // int8→int4 不是隐式转换，函数解析会以
+                                             // 42883「函数不存在」失败（F-80）
 
 const GET_WINDOW_STMT: &str = "select id, state from platform_core.migration_windows where id = $1";
